fix(cli): 🐛 template repo builder

+++ bin/build.js
@@ -35,8 +35,8 @@ logger.info('', {
 
 if (process.argv.includes('info')) process.exit(0);
 
-if (process.argv.includes('clean-template')) {
-  fs.copySync(`${basePath}/images`, `./images`);
+if (process.argv.includes('clean')) {
+  if (fs.existsSync(`${basePath}/images`)) fs.copySync(`${basePath}/images`, `./images`);
   shellExec(`cd ${basePath} && git checkout .`);
   shellExec(`cd ${basePath} && git clean -f -d`);
   process.exit(0);
